working on typing issue

+++ src/app/api/project/route.tsx
@@ -36,12 +36,14 @@ export async function GET() {
 
 export async function DELETE(
     request: NextApiRequest & { nextUrl: URL }
-): Promise<Response> {
-    const id = request.nextUrl.searchParams.get("id");
-    await connectMongoDB();
-    await Project.findByIdAndDelete(id);
-    return NextResponse.json(
-        { message: "Project deleted successfully" },
-        { status: 200 }
-    );
+) {
+    if (request.nextUrl) {
+        const id = request.nextUrl.searchParams.get("id");
+        await connectMongoDB();
+        await Project.findByIdAndDelete(id);
+        return NextResponse.json(
+            { message: "Project deleted successfully" },
+            { status: 200 }
+        );
+    }
 }
